Add basic validation and authorize post/put routes

+++ server/controllers/FilmRental/filmRental.ts
@@ -6,13 +6,18 @@ import { Request, Response } from 'express';
 
 export const GetAll = asyncHandler(async(req: Request, res: Response) => {     
      const filmRentals = await FilmRental.find();
-     res.json(filmRentals);
+     res.status(200).json(filmRentals);
 })
 
 export const GetById = asyncHandler(async(req: Request, res: Response) => {
     const { id } = req.params;
+    const filmExists = await FilmRental.findById( id )
+    if (!filmExists) {
+        res.status(400)
+        throw new Error('film not exist')
+    }
     const filmRental = await FilmRental.findById(id);
-     res.json(filmRental);
+     res.status(200).json(filmRental);
 })
 
 export const Add = asyncHandler(async(req: Request, res: Response) => {
@@ -25,20 +30,35 @@ export const Add = asyncHandler(async(req: Request, res: Response) => {
         rentDate,
         createdAt
     });
+    res.status(200).json({
+        message: "Added succesfuly" 
+    })
 })
 
 export const Edit = asyncHandler(async(req: Request, res: Response) => {
     const { id } = req.params;
+    const filmExists = await FilmRental.findById( id )
+    if (!filmExists) {
+        res.status(400)
+        throw new Error('film not exist')
+    }
     const { firstName, lastName, filmName, rentDate, createdAt } = req.body;
     
     const updatedFilmRental = await FilmRental.findByIdAndUpdate(id, {firstName, lastName, filmName, rentDate, createdAt});
-    res.json(updatedFilmRental);
+    res.status(200).json(updatedFilmRental);
 })
 
 export const Delete = asyncHandler(async(req: Request, res: Response) => {
     const { id } = req.params;
-    const filmRental = await FilmRental.findByIdAndRemove( id );
-    res.json({ message: 'Rental deleted' });
+    const filmExists = await FilmRental.findById( id )
+    console.log(filmExists);
+    if (!filmExists) {
+        res.status(400)
+        throw new Error('film not exist')
+    }
+
+    await FilmRental.findByIdAndRemove( id );
+    res.status(200).json({ message: 'Rental deleted' });
 })
